backend/requirements.txt


# ── Web Framework ────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ── Data Science ─────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
joblib==1.4.2

# ── Deep Learning ────────────────────────────────────────
torch==2.3.0
# torch-cpu only (faster install, enough for LSTM inference)
# For GPU training: pip install torch --index-url https://download.pytorch.org/whl/cu118

# ── Geospatial ───────────────────────────────────────────
geopandas==0.14.4
rasterio==1.3.10
shapely==2.0.4
pyproj==3.6.1
fiona==1.9.6

# ── HTTP / APIs ──────────────────────────────────────────
requests==2.32.3
httpx==0.27.0
aiohttp==3.9.5

# ── Utilities ────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.3
pydantic-settings==2.3.3
loguru==0.7.2

# ── Dev / Testing ────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
black==24.4.2
